+++ backend/src/main/java/com/example/Main/domain/Post/Comment/controller/ApiV1PostCommentController.java
@@ -14,11 +14,9 @@ import com.example.Main.domain.Post.entity.Post;
 import com.example.Main.domain.Post.service.PostService;
 import com.example.Main.global.ErrorMessages.ErrorMessages;
 import com.example.Main.global.RsData.RsData;
-import com.example.Main.global.Security.SecurityMember;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
@@ -26,7 +24,7 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(value = "/api/v1/post/{postId}/comments")
+@RequestMapping(value = "/api/v1/posts/{postId}/comments")
 public class ApiV1PostCommentController {
 
     private final PostCommentService commentService;
@@ -165,30 +163,6 @@ public class ApiV1PostCommentController {
 
         commentService.deleteComment(commentId);
         return RsData.of("200", "%d 번 댓글 삭제 성공".formatted(commentId), null);
-    }
-
-    // 관리자용 댓글 삭제
-    @PreAuthorize("hasRole('ADMIN')")
-    @DeleteMapping("/admin/{commentId}")
-    public RsData<String> deleteCommentByAdmin(@PathVariable("commentId") Long commentId, @AuthenticationPrincipal SecurityMember loggedInUser) {
-        if (loggedInUser == null) {
-            return RsData.of("401", ErrorMessages.UNAUTHORIZED, null);
-        }
-
-        String role = loggedInUser.getAuthorities().toString();
-        if (!role.contains("ROLE_ADMIN")) {
-            return RsData.of("403", ErrorMessages.ONLY_ADMIN, null);
-        }
-
-        PostComment comment = commentService.getComment(commentId).orElse(null);
-
-        if (comment == null) {
-            return RsData.of("404", ErrorMessages.COMMENT_NOT_FOUND, null);
-        }
-
-        commentService.deleteCommentByAdmin(commentId);
-
-        return RsData.of("200", "%d 번 댓글 삭제 성공 (관리자 삭제)".formatted(commentId), null);
     }
 
     // 게시글 댓글 좋아요
